[11.0.0] Dropped utils, changed integer to long, discount/gift effects, okhttp3 version (#105)
* [11.0.0] Dropped utils, changed integer to long, discount/gift effects, okhttp3 version

* CR fixes

* Readme link fix

+++ src/main/java/io/voucherify/client/model/customer/response/CustomerLoyalty.java
@@ -18,7 +18,7 @@ public class CustomerLoyalty {
   private Long points;
 
   @JsonProperty("referred_customers")
-  private Integer referredCustomers;
+  private Long referredCustomers;
 
   private Map<String, CustomerCampaignLoyalty> campaigns;
 }
