+++ src/main/java/io/voucherify/client/model/promotion/response/ListPromotionTiersResponse.java
@@ -17,7 +17,7 @@ public class ListPromotionTiersResponse {
 
   private String object;
 
-  private Integer total;
+  private Long total;
 
   @JsonProperty("data_ref")
   private String dataRef;
